+++ lib/app/ui/components/cache_image.dart
@@ -1,3 +1,4 @@
+import 'package:cm_app/more_libs/setting_v2.8.3/setting.dart';
 import 'package:flutter/material.dart';
 import 'package:t_widgets/t_widgets.dart';
 
@@ -7,6 +8,15 @@ class CacheImage extends StatelessWidget {
   const CacheImage({super.key, required this.url, this.cachePath});
   @override
   Widget build(BuildContext context) {
-    return TCacheImage(url: url);
+    if (Setting.getAppConfig.isUseCacheImageWidget) {
+      return TCacheImage(
+        url: url,
+        errorBuilder: (context, error, stackTrace) {
+          debugPrint('url: $url\n[error]: ${error.toString()}');
+          return TImage(source: url);
+        },
+      );
+    }
+    return TImage(source: url);
   }
 }