+++ lib/app/ui/details/movie_download_list_page.dart
@@ -21,14 +21,26 @@ class _MovieDownloadListPageState extends State<MovieDownloadListPage> {
 
   @override
   Widget build(BuildContext context) {
-    return ListView.builder(
-      controller: controller,
-      shrinkWrap: true,
-      physics: NeverScrollableScrollPhysics(),
-      // primary: false, // controller reuse မဖြစ်စေဖို့
-      itemCount: widget.list.length,
-      itemBuilder: (context, index) => _getlistItem(widget.list[index]),
+    return CustomScrollView(
+      slivers: [
+        SliverList.builder(
+          // controller: controller,
+          // shrinkWrap: true,
+          // physics: NeverScrollableScrollPhysics(),
+          // primary: false, // controller reuse မဖြစ်စေဖို့
+          itemCount: widget.list.length,
+          itemBuilder: (context, index) => _getlistItem(widget.list[index]),
+        ),
+      ],
     );
+    // return ListView.builder(
+    //   controller: controller,
+    //   shrinkWrap: true,
+    //   physics: NeverScrollableScrollPhysics(),
+    //   // primary: false, // controller reuse မဖြစ်စေဖို့
+    //   itemCount: widget.list.length,
+    //   itemBuilder: (context, index) => _getlistItem(widget.list[index]),
+    // );
   }
 
   Widget _getlistItem(MovieDownloadLink link) {